Add goofy tutorial music
