+++ weave/Assets/Music/tutor.mp3.import
@@ -1,0 +1,19 @@
+[remap]
+
+importer="mp3"
+type="AudioStreamMP3"
+uid="uid://7fu6ktqrmcnc"
+path="res://.godot/imported/tutor.mp3-ebd3be676e89764e4ba7907c4f048f30.mp3str"
+
+[deps]
+
+source_file="res://Assets/Music/tutor.mp3"
+dest_files=["res://.godot/imported/tutor.mp3-ebd3be676e89764e4ba7907c4f048f30.mp3str"]
+
+[params]
+
+loop=true
+loop_offset=0.0
+bpm=0.0
+beat_count=0
+bar_beats=4
